awesome-ai-builds/intermediate/code-reviewer-bot/requirements.txt
fastapi==0.115.6
httpx==0.28.1
openai==1.58.1
PyGithub==2.5.0
pytest==8.3.4
python-dotenv==1.0.1
google-generativeai==0.8.3
uvicorn==0.32.1
